feat(sync, customer): Implement data synchronization and customer management
This commit introduces a comprehensive data synchronization system for core entities and adds new customer management functionality. It includes a major refactoring of the data sources, models, and BLoC layers to support these features and improve architectural consistency.

New Features:
- **Data Synchronization**: Adds dedicated BLoCs (`SyncDiscountBloc`, `SyncTaxBloc`, `SyncServiceChargeBloc`) and modifies the `SyncDataPage` to allow users to sync specific data types from the server.
- **Customer Management**: Implements a new `CustomerBloc` to handle fetching and managing customer data.

Refactoring & Improvements:
- **Data Layer**: Updated most remote and local datasources for consistency and to support new models. Introduced `OrderResponseModel` and modified `ProductResponseModel`, `OrderRequestModel`, and others to align with backend API changes.
- **State Management**: Heavily refactored `OrderBloc` to handle more complex state logic. Updated `ProductBloc`, `CategoryBloc`, and `CheckoutBloc` to align with new data structures.
- **Dependency Injection**: Integrated all new BLoCs into `main.dart` to make them available throughout the widget tree.
- **UI/UX**: Modified various pages and widgets (Home, Order, History, Settings) to consume the new state and provide a consistent user experience.
- **Tooling**: Updated the Postman collection to reflect the latest API endpoints and request/response structures.

+++ lib/data/models/order_item_model.dart
@@ -18,7 +18,7 @@ class OrderItem extends Equatable {
     this.updatedAt,
   });
 
-  int get totalPrice => quantity * product.price;
+  int get totalPrice => quantity * product.price.toInt();
 
   @override
   List<Object?> get props => [
